Convert temperature from a quantity to float if it's a quantity.
Fix case when OBC2 GB model is used and the salt concentration is non-zero. When
the salt concentration is non-zero, the customgbforces implementation has to be
used (since the native kernel does not support salt concentrations).  However,
since the native kernel runs ~2x faster on CUDA (and most likely far more
efficiently on CPU and Reference, most likely), it is worth continuing to use
the native kernel when possible.