Add a usesPeriodicBoundaryConditions() method to all of the Force classes and the System. The System inspects all of its forces and calls usePeriodicBoundaryConditions -- it reports 'true' if at least one of its Forces returns 'true' for usesPeriodicBoundaryConditions.
The default Force implementation is to raise an exception (NotImplementedError).
The System implementation will return true if at least one of its Forces returns
'true' for usesPeriodicBoundaryConditions, _regardless_ of whether or not one or
more of the forces does not implement the method. If at least one of the forces
does not implement usesPeriodicBoundaryConditions and all forces return false,
an OpenMMException is thrown.