Fix Gromacs #define replacement and add support for Gromacs 5
Gromacs uses the preprocessor to replace previously #define'd tokens in strings.
This is used, for instance, to override certain parameters without having to
introduce new file types.  This is how Gromacs implemented ff99SB-ILDN (i.e., by
simply replacing the chi torsion for isoleucine, leucine, aspartate, and
asparagine with a define token).  So OpenMM *was* missing some in-line
parameters that were defined.

The immediate consequence I see is that whenever someone tried to run
ff99SB-ILDN with OpenMM, what they were *really* getting was ff99SB. Not sure
where else, if anywhere, this kind of thing was happening.