Added spherical harmonic  multipoles.
o Permanent and induced moment energies, forces and torques are implemented, for PME direct space.
o The code is activated by a preprocessor macro at the top of AmoebaReferenceMultipoleForce.h, and can be completely disabled.
o Currently, the Cartesian momenents are stored alongside the spherical harmonic moments, but these can eventually be removed when the reciprocal and self terms have been adapted to use spherical harmonics (in progress) and the non-PME routines have been coded.