Legalese at top of PDF, bottom of HTML
Depending on whether we want to build HTML docs or a PDF, we want the legalese
either at the end or the beginning of the index page. There is no way that *I*
know of to do this with a single index.rst file, so have conf.py automatically
build the index.rst file, and inspect the command-line to see if it's building
an HTML or a PDF.  Then construct the index.rst file in the desired order for
the type of document that's being generated.