Add a test for AmoebaAngleForce. There's a problem here, though, since the input angles are in degrees (unlike *every* other angle force out there), and stripUnits automatically reduces all angles to radians if they come in with units.
The approach here is to *slightly* change the API, so that
AmoebaAngleForce.addAngle interprets input angles <2*pi as radians, and >2*pi as
degrees. This is heuristic, but should work in every case out in the wild so
far.

I've also updated the documentation to reflect this behavior, and fixed the
units attached to the return value of AmoebaAngleForce.getAngleParameters() to
return degrees instead of radians.