BUGFIX: Fixed PME energies for the QI code.
The QI direct space energy uses the energy functional

    U = 1/4 (E_p µ_d + E_d µ_p)

while the code it replaced uses the equivalent form

    U = 1/2 E_p µ_d

Therefore the reciprocal and self terms have been adapted to use the more symmetric form.  The bad code yielded incorrect energies when PME is used AND µ_d and µ_p are different.  The forces and torques were not affected.