Enable testing of OpenCL platform on Travis-CI
This patch creates a new build in the Travis-CI matrix that compiles and
runs the OpenCL tests using the CPU OpenCL device (AMD drivers). Because
not all of the tests can complete in the allotted 50 minutes, the ctest
invocation is changed to run them in a random order.

This also adds a new executable, TestOpenCLDeviceQuery, that prints
out information about all of the OpenCL devices on the system.