Fix #1213 addInteractionGroup should accept `set` values as parameters.
apparently SWIG typemaps maps Python's list but not set with std::set.
Instead, Python's `set` is mapped with `std::unordered_set`.

The original code of `swigInputBuilder.py` is kindly in a mess. Too much
lines in a single method `writeMethods` which causes debug becomes
difficult. Also there are a lot of warnings about
[PEP8](https://www.python.org/dev/peps/pep-0008/).

A refactoring in the future may be a good idea.