Adds residue name and template name matching
In original version, when a template for a residue is searched, first
match for the elements-bonds graph is selected. This is problematic,
when one has ffxml with separate residues for different isotopologues,
where the elements are the same, but atom masses differ. If a file with
different isotopologues is read in, all of them will be assigned the
same template, which is an undesired behavior.
This commit adds residue name matching. If the name matching fails, the
template matching is not influenced.

	modified:   wrappers/python/simtk/openmm/app/forcefield.py