Precompute GBNeck terms
The previous version of this code would compute
the values of d0 and m0 for every iteration. This
was unnecessary as these values are fixed at the
start of the calculation by the radii of the particles
involved. This resulted in unnecessary computation
and memory access that dramatically slowed simulations
using the GBn or GBn2 solvation models.

The new version pre-computes the values based on the
radii that are present in the system. Linear interpolation
is used, which is consistent with Amber. The previous
version of OpenMM used cubic, which gives slightly
different results.