Try to make checkpoint reporting atomic
Previously, re-starting a simulation would overwrite the existing checkpoint file *at CheckpointReporter creation time*. If your simulation was killed before you wrote a new checkpoint, you'd be out of luck.

With this change, passing in a file object still works the same way. If you pass a filename, the file object will be opened and closed at the time of checkpointing. If you pass an optional new parameter `tempFile`, we'll write to this file instead, and use `os.replace` to atomically overwrite the previous checkpoint. (this part requires python 3)