Lepton: Add three functions that can be used as activation functions in neural network.
I am implementing the forward propagation of the dense neural network as
a CV in Colvars, and this implementation allows to use a custom
activation function defined by Lepton. However, it's hard to write an
expression doing the same things as ReLU and ELU by combining the
existing functions, so in this commit, three additional functions are
added:

1. select2(x, expr1, expr2): Evaluate expr1 when x>0, otherwise
evaluate expr2. Basically this is a helper operation for ELU and RELU.

2. elu(x, alpha): Exponential linear unit. Evaluate alpha*(exp(x)-1)
when x>0, otherwise evaluate x.

3. relu(x): Rectified linear unit. Evaluate x when x>0, otherwise return
0.