Remove tri-sulfite bond formation. (#3203)
* Remove tri-sulfite bond formation.

Pdbfixer attempts to add di-sulfide bonds to any pair of CYS "S" atoms that are less than 3 A. This procedure fails when multiple such bonds are assigned to a given atom. 

Here we rectify this issue by a greedy algorithm: Atom I is chosen to be bonded to `argmin(distance([:I-1, I])` over all candidate di-sulfide bonds. This procedure will not generally result in a set of minimal distance di-sulfide bonds, but will fix the "tri-sulfide bond" problem. Fixing the more general case would require a solution to global constraint satisfaction problem.

**Legal**
This patch is provided under the CC0 license (https://creativecommons.org/share-your-work/public-domain/cc0/).
This is not an officially-supported Google product. It is provided ‘as-is’ without any warranty of any kind, whether expressed or implied.

* Update topology.py

Remove a redundant comment.