Fix misleading documentation of `Topology.bonds`
`Topology.bonds()` states that iterates over a tuple of `Atom` objects. While this is technically correct, as the `Bond` class is a named tuple, this also hides the fact, that `Bond` objects also provide the `order` and `type` attributes. Therefore, I propose to change the sentence, so that it uses the capitalized *Bonds* indicating that they are actually objects.