Support `auth_seq_id` in `_struct_conn` PDBx section
Both in reading and writing.
For reading, we get residue numbers preferentially from
`auth_seq_id` in the `_atom_site` section, so we should use
the same in `_struct_conn` if it is available.
For writing, we duplicate the residue number in `_struct_conn`
the same way we do so in `_atom_site`.