Args to consider multi template matches unmatched
Add boolean arguments to _getResidueTemplateMatches and
getUnmatchedResidues such that multiple template matches can be considered
unmatched rather than raising an exception. Raising an exception isn't
useful behaviour if you have specified the residueTemplates dictionary
correctly and neither of the existing matches are suitable for the
residue in question. This is useful when checking for residues that need
to have parameters generated for them. However, keep the default
behaviour as it was as seeing an exception may be useful in other
workflows such as trying to diagnose a parameterisation failure.

(SFT-31575)